requirements.txt
numpy==1.21.6
pandas==1.5.2
openai==0.27.5
mysql==0.0.3
mysqlclient==2.1.1
mysql-connector-python-rf==2.2.2
python-dotenv==1.0.0
flask==2.2.2
SQLAlchemy==2.0.12
llama-index==0.6.1
transformers==4.28.1